Add systemd service for platforms with systemd
The opentsdb_instance resource was only providing support to sysvinit
system.

This commit adds a systemd based service unit template for newer debian
and ubuntu systems. There is also a new java_home attribute to the
resource to set the JAVA_HOME env in the service definitions.

It has been tested against Chef 13.